Let backup_ip checks be coherent
Use 'backup_ip is defined and backup_ip != ""' everywhere. It is already
used in the dns-server-bind playbook.

The external-ip-type common playbook sets the backup_ip fact to the
empty string if `network.auto_detect_ip` is false and
`network.backup_ip` is null, or if `network.auto_detect_ip` is true and
the obtained `backup_ip` is different from the external_ip.

The checks cannot reliably be made on `network.backup_ip`.

Use `backup_ip` instead of `network.backup_ip` where relevant.