Fix certificates domain alias handling
Avoid certificates getting a domain_alias from a previous run of the
certificate role.

The `domain_alias_cmd` fact was left untouched if the requested domain
alias was undefined or empty (eg. undefined domain_alias in
cert-imap.yml), which could lead to the use of a
`domain_alias_cmd` value from a previous run of the role in another
playbook (eg. smtp2 from cert-smtp.yml).

Remove the use of the `domain_alias_cmd` fact and set the command line
option conditionally with templating.