Use only TLS ports with Dovecot
Keep the `imap` listener on port `143` for the needs of the webmails
through `imapproxy` on localhost.

Disable the `pop3` listener on port 110, even if `mail.pop3` is enabled.

Close the firewall ports for `imap` and `pop3`.

Announce only `imaps`, and `pop3s` if POP3 is enabled:

- with the DNS SRV records;

- in `autoconfig` and `autodiscover`.

Update the tests for the DNS SRV records.