Certificates: do not redirect to www
In nginx configurations for certificate requests, always redirect any
request not related to certificate requests to:
```
rewrite ^ https://$server_name$request_uri? permanent;
```

Do not redirect the APEX domain to `wwww.` since we do not know if it
exists. Let the server configuration on `https` for `$server_name` do
it for us if needed.

It might give two redirections in a row, but since redirections are
permanent and HSTS is configured, it should be a limited case.