Add a default catchall Nginx server block
Configure a catchall server as default server for HTTP (80) and HTTPS
(443) ports. This server block catches all requests done on an
unconfigured server name or on one of the IP address directly.

Close all connections that get to this server block (`return 444`).

Use the apex domain certificate on the HTTPS port.

Based on :

- https://www.nginx.com/resources/wiki/start/topics/examples/server_blocks/#a-default-catch-all-server-block

- https://nginx.org/en/docs/http/server_names.html#miscellaneous_names