Set backup_ip_type only if backup_ip is defined
`network.backup_ip` is set to `None` (`~`) in the default configuration.

If it is not configured in the user configuration or if it is set to the
empty string, `backup_ip` and `backup_ip_type` will be undefined.

Test if `backup_ip_type` is defined when trying to determine if IPv6 is
used.

The usage of `backup_ip` and `backup_ip_type` is for now guarded
elsewhere with (`backup_ip is defined and backup_ip != ""`).