Set the backup_ip fact in any case
A different task condition between the task that set `backup_ip` and the
one getting the IP type led to an undefined `backup_ip` error. The tasks
conditions should be the same.

`network.backup_ip` is set to `None` (`~`) in the default configuration.
If it is not configured in the user configuration, it will give an empty
string for `backup_ip`, which is ok since it is guarded that way
everywhere needed (`backup_ip is defined and backup_ip != ""`).

Remove the conditions on `network.backup_ip` and set `backup_ip` and
`backup_ip_type` in any case.