Use the ipv6_used fact in system-prepare
Add the `external-ip-type` common role as a dependency of the
`system-prepare` role.

Use the `ipv6_used` fact established in the `external-ip-type` role to
decide if IPv6 support on the server is to be disabled.

Allow for reconfiguration by setting the sysctl parameter value (`0` or
`1`) corresponding to the boolean value of `ipv6_used`.