Remove condition to determine if backup_ip is IPv6
The default value of `None` would not satisfy the condition and the
`backup_ip_ipv6` fact would not be set although it is needed by the
following task condition.

Without condition, if the `network.backup_ip` is not configured by the
user, it will have a value of `None` which will fail the `ipv6` test, so
`backup_ip_ipv6` will be `false` as expected.