borgbackup: separate server from users repos
Use a '@server' directory inside the backup location to store the server
backup repository.

The backup directories generated for the users (to be used with
backup-server) are then at the same level. If the backup location is
listed it will look like :
```
@server/
uid1/
uid2/
```

Avoid naming conflicts by using `@` which cannot be used in a UID.