borgbackup: mountable filesystems
Bring back support for CIFS filesystems. It was lost during refactoring.

Separate `cifs` and `sshfs` cases where the target directory is directly
mounted on the `mountPath` from `s3fs` and `usb` where the root of the
filesystem is mounted on the `mountPath`.

Bring back the location path for `s3fs` and `usb` filesystems. For these
filesystems, the URL only need to specify the path on the remote mount
system. No reference to `/mnt/backup/<location.name>` is needed. Update
the doc accordingly.