import-accounts: use IMAPS with TLSv1.2 in isync
Configure the TLS protocol version to TLSv1.2. Isync uses TLSv1 by
default. Configuring the protocol version is needed to access the local
IMAP server configured with Mozilla's Intermediate profile using TLSv1.2
only.

Use IMAPS to connect to the local store. The connection to the
localstore uses STARTTLS by default. It uses TLS anyway, so make it
explicit and directly use IMAPS on the port 993 with TLSv1.2.