Make TLS parameters configurable
Factor the main TLS parameters as variables in `config/defaults.yml`.

Add a task in the `common/roles/load-defaults` role filtering out the
`TLSv1.3` protocol version for Debian 9 (Stretch). The OpenSSL version
packaged in Stretch (1.1.0l) does not support TLSv1.3.

Filter out the `SSLv2` version in Dovecot. It does not support this
parameter.