sogo: Create SOGo's sieve filter file
Allow for the configuration of sieve filters through the SOGo user
interface. If the `sogo.sieve` file is absent, SOGo fails to add the
rules as sieve filters.

SOGo seems to correctly create the file when configuring vacation
filters. Having both might give conflicts.

See:
https://lists.inverse.ca/sogo/arc/users/2016-06/msg00198.html

And for the full discussion:
https://lists.inverse.ca/sogo/arc/users/2017-05/msg00073.html