Add Pebble as local ACME server option for devel
Add a docker-compose file for Pebble. Pebble is a test server for the
ACME protocol (Letsencrypt). The Pebble Challenge Test Server is also
used as configurable DNS server for Pebble (ie used as TLD name server).
https://github.com/letsencrypt/pebble
https://github.com/letsencrypt/pebble/blob/master/cmd/pebble-challtestsrv/README.md

Add an option to declare an ACME server for development. Default to
"staging" to keep the current behaviour. Can be set to "pebble".

Install the temporary root CAs generated by Pebble during installation
and tests when needed.

Configure the `external_ip` dynamically as default response from the
Pebble name server.

Do not backup certificates when using Pebble.