Tag certificates role imports with 'certificates'
Allow for complete certificates generation and configuration through the
'certificates' tag.

Tag the certificates role in all playbooks. Through tag inheritance, all
tasks from the role, including on the role dependencies, are tagged.

Tag the imports of the `cert-perms` role with 'certificates'.

Use `import_role` for `gencert.yml` in the `certificates` tasks to get
tag inheritance.

Use `apply` to specify the `certificates` tag on the dynamic
`include_role` needed for the loop in the `extra-certs` role.