certificates-all: Set keys access with acl module
Restrict any existing private keys access rights (from `644` →
`root:root 600`) before any certificate is issued. Certbot uses these
rights on renewal (among other things) to determine the rights for a new
key.

Use the acl module to restrict standard access rights without
recalculating the ACL mask. This is a bit more verbose but avoids a
potential failure to restart on renewal (eg slapd) and having to wait
for cert-perms to fix the ACL rules.

Do not try to restrict the access rights on the certificates: it is not
really needed. Default access rights set by certbot for certificates are
`644`.

Overall, this does not seem to significantly change the execution time.
The tasks run only for private keys found with `644` access rights.