certs: Align with certbot access rights
Set the access rights on the archive and live directories as certbot
would:
- set archive and live `700` ;
- let certbot create per fqdn directory (with `755`) ;
- let certbot create certificates with `644` (under the `755` dirs),
private keys `600` for certbot >= 0.29 (buster).

Avoid any ACL mask reset:
- use the acl module without mask recalculation to set access rights ;
- do not set access rights prematurely on archive and live directories
in system-prepare.

Get back to setting ACL rules from the archive and live directories and
down. Only allow for traversal of the directories (`x`), not listing.

Explicitly set default ACL rules on per fqdn directories. This ensures
proper ACL rules and mask on all files for a user with access to the
private key (eg openldap). A user with access only to the certificates
will need a call to `setfacl` in a renewal hook (cf dovecot).