extra-certs: use the certificates-all role
Migrate the generation of the certificates configured with `extra_certs`
to use the `certificates-all` role.

Remove the vars from the `certificates-all` role to specify the
certificates list as role parameter in the main playbook. This allows to
use another certificates list, ie an adapted `extra_certs`, for
additional certificates. It was not possible for some reason to
overwrite the role's vars with role parameters in the main playbook.

Add an `extra-certs` tag to be able to only generate those certificates
and replace the `extra-certs.yml` playbook.

Remove the `extra-certs.yml` playbook and the `extra-certs` role.