Add an OpenPGP Web Key Directory role
Use both the direct and advanced publishing methods.

- The direct method uses the well-known directory of the base domain.
  Some PGP clients might only support this version.

- The advanced method uses an `openpgpkey` subdomain.

https://tools.ietf.org/html/draft-koch-openpgp-webkey-service

The public keys are expected to be in the server configuration in a
"pgp" object list along with the user's uid.

```
pgp:
  - uid: marie
    ascii_armored_public_key: |
      -----BEGIN PGP PUBLIC KEY BLOCK-----

      […]
      -----END PGP PUBLIC KEY BLOCK-----
```

The server's `gpg` is used to determine the WKD hash for each key, which
is then used to name the published files.